requirements.txt
streamlit==1.30.0
langchain==0.0.330
openai==1.0.0
faiss-cpu==1.7.4
tqdm
PyPDF2
python-multipart
pdfminer.six
pydantic
sentence-transformers==2.2.2
chromadb==0.4.9
